+++ features/login/impl/src/main/kotlin/io/element/android/features/login/impl/root/LoginRootPresenter.kt
@@ -25,7 +25,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import io.element.android.features.login.api.oidc.OidcAction
-import io.element.android.features.login.impl.oidc.web.DefaultOidcActionFlow
+import io.element.android.features.login.impl.oidc.customtab.DefaultOidcActionFlow
 import io.element.android.features.login.impl.util.LoginConstants
 import io.element.android.libraries.architecture.Async
 import io.element.android.libraries.architecture.Presenter
